allow macros to generate definitions etc...
This was mostly done by rewriting `expand-body`,
which now has a second pass scanning for declarations
& definitions *after* macros have been expanded.

This paved the way for a `make-enum` macro which can
generate a new datatype and several objects and functions
related to it (and probably lots of other nice features to come).

A new reflection-related special form `%%symbol` allows macros
to construct symbols from concatenated (symbol|string|int).