reflection using the new metadata.
A couple of new C prims are added that allow some of the
runtime smarts (in places like `dump_object`) to be written
in Irken directly.  The motivator for this was the need to
know the names of uncaught exceptions, a thorn in my side
for many moons.

Eventually it should be possible to extend reflection so that
it can examine any object at runtime, knowing its type; making
things like a debugger finally achievable.