genffi: C parser completely rewritten.
This was done in two stages:
1) the 'antlr2irken' utility translated 'C.g4' from the
   antlr distribution into an irken lexer & grammar.
2) these were manually tweaked (mostly to add the 'lex hack',
   which returns typedef'd names as Typedef terminals rather
   than Identifiers).

All of this difficult work was triggered by the presence of a single
inline function definition in an include file, which threw the old
code off. [you can't skip past the function definition unless you can
parse it!]

Aside from fixing the original problem, this full parser handles
many more difficult cases, and as a bonus we now support enums.
[though that support has not made it fully into the ffi spec language yet]