Change tag size from 8 to 16 bits.
While working on another branch I came across two issues:
1) the number of available polyvariants is quite limited (54!) and
this is too few given their use with exceptions.
2) even though there was a limit, it was not being checked, and this
led to mystery crashes.

The choice of 8-bit tagsize was made in the 32-bit era, and is no
longer practical.

Of _course_, all the shifting & masking was done with literal '8' and
'255' constants, not even a C macro.  So I've made that improvement.
However, changing the tagsize is NOT so straightforward:

1) There are places in the code that rely on pre-encoded constants for
immediates - like the runtime 'cmp' functions - but especially the LLVM
preamble.  There's no facility in the language for those kinds of
constants, so any future change to the tag size should use this diff
as a guideline.  [the nastiest one was a shift of 7 bits caused by an
optimized bit of LLVM in irk_tuple_len]
2) like most major internal changes, there's the "out-on-a-limb"
bootstrapping issue.  Here's how I pulled it off this time:

a) compile the new compiler using the old bytecode compiler:
    irkvm self/compile.scm -b
b) compile again with the new vm (with the new tagsize):
    vm/irkvm self/compile.byc -b
c) bootstrap the C (and llvm) backends from there:
    vm/irkvm self/compile.byc self/compile.scm